AdvancedSampler: Fix empty cache playback
Playing back an empty cache causes a crash. A test for
this seems to have been added right at the beginning of the
main process. Unfortunately this causes the whole main process
to not execute (including the ability to record) unless the cache
is not empty. Move the test to surround only the playback actions.